Keep lifecycle tests out of install consumer
Build the private-header lifecycle coverage only in the repository unit test
target so the installed-package test remains limited to public headers.

Copilot-Session: 2af0dd12-3cd9-4c21-a94c-0d20612e1045
